Move logger to own module and simplify
